Rename file to avoid clippy warning.
