+++ hugr/src/std_extensions/arithmetic/int_ops/const_fold.rs
@@ -1201,4 +1201,4 @@ pub(super) fn set_fold(op: &IntOpDef, def: &mut OpDef) {
 }
 
 #[cfg(test)]
-mod test;
+mod tests;
